# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification. Requests to an unknown route under `/api` return a 404
    response using the `Error` schema with the message `API route not found`.
    Requests with malformed JSON bodies return a 400 response using the `Error`
    schema with the message `Malformed JSON request body`. Unexpected server
    failures return a generic 500 response using the same schema. JSON request
    bodies that exceed the parser's size limit return a 413 response using the
    `Error` schema with the message `Request body too large`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /practice/sessions:
    get:
      operationId: listPracticeSessions
      tags: [practice]
      summary: List a learner's completed practice sessions
      responses:
        "200":
          description: Completed practice session history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeSessionSummary"
        "400":
          description: Learner identity is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createPracticeSession
      tags: [practice]
      summary: Start a learner practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeSessionInput"
      responses:
        "201":
          description: Started practice session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "400":
          description: Invalid practice session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: JSON request body exceeds the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /practice/sessions/latest:
    get:
      operationId: getLatestPracticeSession
      tags: [practice]
      summary: Get the learner's latest completed session
      responses:
        "200":
          description: Latest completed session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "404":
          description: No completed session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /practice/sessions/{id}:
    get:
      operationId: getPracticeSession
      tags: [practice]
      summary: Get a learner practice session with its answers
      parameters:
        - $ref: "#/components/parameters/PracticeSessionId"
      responses:
        "200":
          description: Practice session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "404":
          description: Practice session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: savePracticeAnswer
      tags: [practice]
      summary: Save an answer in a learner practice session
      parameters:
        - $ref: "#/components/parameters/PracticeSessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAnswerInput"
      responses:
        "200":
          description: Saved practice answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAnswer"
        "400":
          description: Invalid practice answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: JSON request body exceeds the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Practice session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /practice/sessions/{id}/complete:
    post:
      operationId: completePracticeSession
      tags: [practice]
      summary: Complete a learner practice session
      description: Completes the session manually, or records time_expired when a timed session's server-side duration has elapsed.
      parameters:
        - $ref: "#/components/parameters/PracticeSessionId"
      responses:
        "200":
          description: Completed practice session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "404":
          description: Practice session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /practice/sessions/{id}/abandon:
    post:
      operationId: abandonPracticeSession
      tags: [practice]
      summary: Abandon an unfinished learner practice session
      parameters:
        - $ref: "#/components/parameters/PracticeSessionId"
      responses:
        "200":
          description: Abandoned practice session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSession"
        "400":
          description: Session cannot be abandoned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Practice session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    PracticeSessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    QuestionDomain:
      type: string
      enum: [People, Process, "Business environment"]
    QuestionApproach:
      type: string
      enum: [agile, predictive, hybrid]
    QuestionDifficulty:
      type: string
      enum: [easy, medium, hard]
    QuestionStatus:
      type: string
      enum: [draft, published, archived]
    Question:
      type: object
      required:
        - id
        - domain
        - topic
        - approach
        - difficulty
        - question
        - translation
        - options
        - correctAnswer
        - explanation
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        domain:
          $ref: "#/components/schemas/QuestionDomain"
        topic:
          type: string
          minLength: 1
        approach:
          $ref: "#/components/schemas/QuestionApproach"
        difficulty:
          $ref: "#/components/schemas/QuestionDifficulty"
        question:
          type: string
          minLength: 1
        translation:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: string
            minLength: 1
        correctAnswer:
          type: integer
          minimum: 0
        explanation:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/QuestionStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PracticeMode:
      type: string
      enum: [quick, domain, mock]
    PracticeSessionStatus:
      type: string
      enum: [in_progress, completed, abandoned]
    PracticeCompletionReason:
      type: string
      enum: [manual, time_expired]
    PracticeSessionInput:
      type: object
      required: [mode, questionIds, timed]
      properties:
        mode:
          $ref: "#/components/schemas/PracticeMode"
        domain:
          type: ["string", "null"]
        approach:
          type: ["string", "null"]
        timed:
          type: boolean
        questionIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
    PracticeAnswerInput:
      type: object
      required: [questionId, selectedAnswer]
      properties:
        questionId:
          type: string
          minLength: 1
        selectedAnswer:
          type: integer
          minimum: 0
    PracticeAnswer:
      type: object
      required: [id, sessionId, questionId, selectedAnswer, isCorrect, answeredAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        questionId:
          type: string
        selectedAnswer:
          type: integer
          minimum: 0
        isCorrect:
          type: boolean
        answeredAt:
          type: string
          format: date-time
    PracticeSession:
      type: object
      required: [id, mode, timed, questionIds, status, completionReason, score, startedAt, completedAt, questions, answers]
      properties:
        id:
          type: string
        mode:
          $ref: "#/components/schemas/PracticeMode"
        domain:
          type: ["string", "null"]
        approach:
          type: ["string", "null"]
        timed:
          type: boolean
        questionIds:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/PracticeSessionStatus"
        completionReason:
          $ref: "#/components/schemas/PracticeCompletionReason"
        score:
          type: ["integer", "null"]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PracticeAnswer"
    PracticeSessionSummary:
      type: object
      required: [id, mode, score, questionCount, completedAt, completionReason]
      properties:
        id:
          type: string
        mode:
          $ref: "#/components/schemas/PracticeMode"
        score:
          type: integer
        questionCount:
          type: integer
        completedAt:
          type: string
          format: date-time
        completionReason:
          $ref: "#/components/schemas/PracticeCompletionReason"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string